+++ src/index.ts
@@ -1,7 +1,9 @@
 import * as parsingUtils from "./helpers/parsingUtils";
+import TarButton from "./components/TarButton.vue";
 import TarCheckbox from "./components/TarCheckbox.vue";
 import TarSpinner from "./components/TarSpinner.vue";
 
+export * from "./types/TarButton";
 export * from "./types/TarCheckbox";
 export * from "./types/TarSpinner";
-export { TarCheckbox, TarSpinner, parsingUtils };
+export { TarButton, TarCheckbox, TarSpinner, parsingUtils };