+++ CHANGELOG.md
@@ -10,5 +10,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Created and configured the helper functions & component library project.
+- Implemented a Button component.
 - Implemented a Checkbox component.
 - Implemented a Spinner compoment.